requirements.txt
asgiref==3.11.0
Django==5.2.8
django-filter==25.2
sqlparse==0.5.4
tzdata==2025.2
